Rename client to generator
