+++ lib/routes_to_swagger_docs/generator.rb
@@ -1,8 +1,8 @@
 require 'fileutils'
 require_relative 'routing/parser'
 require_relative 'schema/v3/openapi_object'
-require_relative 'client/base_client'
-require_relative 'client/doc_client'
+require_relative 'generator/base_generator'
+require_relative 'generator/doc_generator'
 
 module RoutesToSwaggerDocs
   class Generator < BaseGenerator
