+++ lib/routes_to_swagger_docs/generator/base_generator.rb
