+++ lib/routes_to_swagger_docs/generator/doc_generator.rb
