+++ lib/routes_to_swagger_docs/generator/path_generator.rb
