+++ lib/routes_to_swagger_docs/generator/schema_generator.rb
